Implement warning for unused dependencies. This will print a diagnostic for crates which are mentioned as `--extern` arguments on the command line, but are never referenced from the source.
This diagnostic is controlled by `-Wunused-crate-deps` or
`#![warn(unused_crate_deps)]` and is "allow" by default.

There are cases where certain crates need to be linked in but are not
directly referenced - for example if they are providing symbols for C
linkage. In this case the warning can be suppressed with
`use needed_crate as _;`.

The diagnostics will name the crate and the source file of its
root module.  Ideally, however, it would point to the place where the
dependency is defined (in Cargo.toml, for example) to allow the problem
to be solved directly.  That still needs some design work and will be
in a later PR.

Resolves issue #57274